Relocate SleepData processing out of FitbitClient and into an extensions class Relocate the associated unit tests into a new test class
